+++ crawler/src/bin/manual.ts
@@ -1,8 +1,8 @@
 import logging from "../lib/logging";
 
 import InstanceQueue from "../queue/instance";
-import CommunityQueue from "../queue/community";
-import SingleCommunityQueue from "../queue/check_comm";
+import CommunityQueue from "../queue/community_list";
+import SingleCommunityQueue from "../queue/community_single";
 import KBinQueue from "../queue/kbin";
 
 export default async function runManualWorker(
